Add Business Logic Retake Exam 15 August 2019  12/13 Tests

+++ CSharp OOP/Exams/Retake Exam - 15 August 2019/01. Structure_Skeleton-1/Models/Astronauts/Biologist.cs
@@ -4,7 +4,7 @@
 
     public class Biologist : Astronaut, IAstronaut
     {
-        private const double DefaultOxygenQuantity = 70.0;
+        private const int DefaultOxygenQuantity = 70;
 
         public Biologist(string name) 
             : base(name, DefaultOxygenQuantity)
